requirements.txt
requests~=2.32.4
torch~=2.7.1
onnxruntime-gpu~=1.22.0
onnx~=1.18.0
transformers~=4.41.0
sentencepiece~=0.2.0
